Wait for ovn at migration
Till now, the network-vif-plugged event was send to nova as soon the
LSP got updated at northbound ovsdb. Than nova starts moving the VM.
But with big environments and slow northd the VM is moved before the
network is prepared at destination hypervisor.

So skip the fake vif-interface-plugged event at
update_port_postcommit() at a migration.

At PortBindingChassisUpdateEvent check for len(req_chassis) > 1, so we
have a migration and southbound ovsdb has the change. If that's the
case, we process the event, so nova get's an network-vif-plugged event
and migration continues.
additional_chassis is set later, but nova waits for the event before
attaching the tap interface. Also based on [1] it would be better to
wait for additional_chassis entry, as ovn-controller than has the
change, if ovn-controller is down, the migration doesn't start. So we
would catch ovn-controller going down between migration API call and
wait for network setup, but if ovn-controller goes down after waiting
for network, we migrate to a host with no network anyway.

[1] https://ihar.dev/posts/ovn-chassis-binding-walkthru/#multichassis-port-bindings

Closes-Bug: #2069718
Change-Id: Ib79dff034298902dd4b12ffd6889240e7f2e2724